refactor(airflow): make the loader DAG canonical (load_people_api), mark legacy as _old
The train-deployment DAG fully replaces the legacy district/stats loader, so drop the _v2
suffix: new DAG is load_people_api (dag_id load_people_api); the legacy DAG becomes
load_people_api_old (file + dag_id). At prod merge this swaps the canonical dag_id to the new
loader — gate the merge on cutover readiness (DATA-1855).
