+++ airflow/astro/dags/load_people_api.py
@@ -1,223 +1,109 @@
-"""
-## Load People-API Tables to PostgreSQL
+"""People-API voter refresh (DATA-1913).
 
-Reads dbt models from Databricks and upserts rows into the people-api
-PostgreSQL database via an SSH tunnel through the bastion host.
+Thin sequencer over the loader CLI (installed on the worker image). Each step is a
+BashOperator running `loader <step> --date {{ ds_nodash }}`; parallelism lives in the
+loader, and inter-step state flows through the loader's S3 manifests (no Airflow xcom).
 
-### Tables Loaded:
-1. **load_districts** — `m_people_api__district` → `"District"`
-   (parent table, must run first due to foreign key constraints)
-2. **load_district_stats** — `m_people_api__districtstats` → `"DistrictStats"`
-   (has FK to District)
-
-### Connections (set in Astro Environment Manager):
-- `databricks` / `databricks_dev` (Generic) — Databricks OAuth M2M credentials
-- `gp_bastion_host` (SSH) — bastion host for tunneling to PostgreSQL
-- `people_api_db` (Postgres) — people-api database credentials
-
-### Variables (set in Astro Environment Manager):
-- `databricks_conn_id` — selects Databricks connection
-  (e.g., `databricks_dev` in dev, `databricks` in prod)
-- `databricks_catalog` — Databricks catalog name (e.g., `goodparty_data_catalog`)
-- `databricks_dbt_schema` — Databricks schema where dbt models live
-  (e.g., `dbt` in prod, `dbt_staging` in dev)
-- `people_api_schema` — PostgreSQL schema name for people-api tables
+Postgres is reached via the gp_bastion_host SSH tunnel — the loader's LOADER_BASTION_*
+env vars (Plan A) are populated from the gp_bastion_host connection. All LOADER_* env
+values are Jinja templates resolved at task runtime, so DAG parsing never touches the
+metastore.
 """
 
-import json
-import logging
+from __future__ import annotations
 
-from airflow.sdk import Param, Variable, dag, get_current_context, task
-from include.custom_functions.databricks_utils import read_databricks_table
-from include.custom_functions.postgres_utils import (
-    get_max_updated_at,
-    get_postgres_via_ssh,
-    upsert_rows,
-)
-from pendulum import datetime, duration
-from psycopg2.extras import Json
+import os
 
-t_log = logging.getLogger("airflow.task")
+from airflow.providers.standard.operators.bash import BashOperator
+from airflow.sdk import dag
+from cosmos import ProfileConfig
+from cosmos.operators.local import DbtTestLocalOperator
+from cosmos.profiles import DatabricksOauthProfileMapping
+from pendulum import datetime as pendulum_datetime
+from pendulum import duration
 
 
-# The dbt SQL model defines struct fields as lowercase, but the API expects
-# camelCase. Remap the two affected keys on the way into PostgreSQL.
-_BUCKET_KEY_MAP = {
-    "presenceofchildren": "presenceOfChildren",
-    "estimatedincomerange": "estimatedIncomeRange",
-}
+def _loader_env() -> dict[str, str]:
+    """Bastion fields for the loader's SSH tunnel, from the gp_bastion_host connection.
 
-DISTRICT_COLUMNS = [
-    "id",
-    "created_at",
-    "updated_at",
-    "type",
-    "name",
-    "state",
-]
+    Templated so they resolve at task runtime (not DAG parse — parse must not touch the
+    metastore). The rest of the loader's config (LOADER_S3_BUCKET, LOADER_S3_IMPORT_ROLE_ARN,
+    LOADER_AWS_ACCOUNT_ID, LOADER_DATABRICKS_WAREHOUSE_ID, VPC/subnet/SG/KMS) is set as
+    deployment env vars in the Astro Environment Manager and reaches the `loader` subprocess
+    via append_env=True — consistent with the loader's native env-var config and with the
+    Cosmos gate's os.environ reads below.
+    """
+    return {
+        "LOADER_BASTION_HOST": "{{ conn.gp_bastion_host.host }}",
+        "LOADER_BASTION_PORT": "{{ conn.gp_bastion_host.port or 22 }}",
+        "LOADER_BASTION_USER": "{{ conn.gp_bastion_host.login }}",
+        "LOADER_BASTION_PRIVATE_KEY": "{{ conn.gp_bastion_host.extra_dejson.get('private_key', '') }}",
+    }
 
-DISTRICT_STATS_COLUMNS = [
-    "district_id",
-    "updated_at",
-    "total_constituents",
-    "total_constituents_with_cell_phone",
-    "buckets",
-]
+
+def _step(task_id: str, subcommand: str) -> BashOperator:
+    return BashOperator(
+        task_id=task_id,
+        bash_command=f"loader {subcommand} --date {{{{ ds_nodash }}}}",
+        env=_loader_env(),
+        append_env=True,
+    )
+
+
+# dbt project shipped in the Astro image; override via env if the build places it elsewhere.
+_DBT_PROJECT_DIR = os.environ.get("LOADER_DBT_PROJECT_DIR", "/usr/local/airflow/dbt/project")
+
+
+def _voter_gate_profile() -> ProfileConfig:
+    """dbt profile for the voter gate, built from the `databricks` (OAuth M2M) connection.
+
+    host / client_id / client_secret come from the connection; schema and http_path (the SQL
+    warehouse) come from deployment env vars, read at parse time because the ProfileConfig is
+    constructed during DAG parsing (Airflow Variables would require the metastore).
+    """
+    warehouse_id = os.environ.get("LOADER_DATABRICKS_WAREHOUSE_ID", "")
+    return ProfileConfig(
+        profile_name="default",
+        target_name="loader",
+        profile_mapping=DatabricksOauthProfileMapping(
+            conn_id="databricks",
+            profile_args={
+                "schema": os.environ.get("LOADER_DBT_SCHEMA", "dbt"),
+                "http_path": f"/sql/1.0/warehouses/{warehouse_id}",
+            },
+        ),
+    )
 
 
 @dag(
-    start_date=datetime(2026, 3, 2),
-    schedule="@daily",
-    max_consecutive_failed_dag_runs=5,
-    max_active_runs=1,
-    doc_md=__doc__,
+    dag_id="load_people_api",
+    schedule="@monthly",
+    start_date=pendulum_datetime(2026, 6, 1, tz="UTC"),
     catchup=False,
-    default_args={
-        "owner": "Data Engineering Team",
-        "retries": 3,
-        "retry_delay": duration(seconds=30),
-    },
-    tags=["people_api", "postgres"],
-    is_paused_upon_creation=True,
-    params={
-        "full_reload": Param(
-            False,
-            type="boolean",
-            description="Skip the updated_at watermark and reload ALL tables from Databricks.",
-        ),
-        "full_reload_districts": Param(
-            False,
-            type="boolean",
-            description="Skip the watermark for District only.",
-        ),
-        "full_reload_district_stats": Param(
-            False,
-            type="boolean",
-            description="Skip the watermark for DistrictStats only.",
-        ),
-    },
+    default_args={"retries": 1, "retry_delay": duration(minutes=5)},
+    tags=["people-api", "loader", "DATA-1640"],
 )
 def load_people_api():
-    @task
-    def load_districts():
-        """Read District from Databricks and upsert into PostgreSQL.
+    inspect_prod = _step("inspect_prod", "inspect-prod")
+    dbt_test_voter_gate = DbtTestLocalOperator(
+        task_id="dbt_test_voter_gate",
+        project_dir=_DBT_PROJECT_DIR,
+        profile_config=_voter_gate_profile(),
+        select=["m_people_api__voter"],  # voter schema tests + the DATA-1906 singular gate test
+    )
+    unload = _step("unload", "unload")
+    provision = _step("provision", "provision")
+    create_schema = _step("create_schema", "create-schema")
+    copy = _step("copy", "copy")
+    build_indexes = _step("build_indexes", "build-indexes")
+    resize = _step("resize", "resize")
+    validate = _step("validate", "validate")
 
-        Loads all districts except federal-level (state='US').
-        Must complete before DistrictStats due to foreign key constraint.
-        """
-        params = get_current_context()["params"]
-        skip_watermark = params.get("full_reload") or params.get("full_reload_districts")
-        catalog = Variable.get("databricks_catalog")
-        schema = Variable.get("databricks_dbt_schema")
-        batch_size = 5000
-
-        pg_schema = Variable.get("people_api_schema")
-
-        watermark = None
-        if not skip_watermark:
-            with get_postgres_via_ssh() as conn:
-                watermark = get_max_updated_at(conn, pg_schema, "District")
-
-        query = (
-            f"SELECT {', '.join(DISTRICT_COLUMNS)} "
-            f"FROM `{catalog}`.`{schema}`.`m_people_api__district` "
-            f"WHERE state != 'US'"
-        )
-        if watermark:
-            query += f" AND updated_at >= '{watermark}'"
-            t_log.info("Incremental load — watermark: %s", watermark)
-        elif skip_watermark:
-            t_log.info("Full reload requested — skipping watermark")
-        else:
-            t_log.info("Initial load — table is empty, no watermark")
-        query += " ORDER BY updated_at ASC"
-
-        t_log.info("Reading from Databricks: %s", query)
-        _col_names, batches = read_databricks_table(query, batch_size=batch_size)
-
-        try:
-            with get_postgres_via_ssh() as conn:
-                total = 0
-                for batch in batches:
-                    total += upsert_rows(
-                        conn=conn,
-                        schema=pg_schema,
-                        table="District",
-                        columns=DISTRICT_COLUMNS,
-                        conflict_columns=["id"],
-                        rows=batch,
-                    )
-        finally:
-            batches.close()
-
-        t_log.info("Upserted %d District rows to PostgreSQL", total)
-
-    @task
-    def load_district_stats():
-        """Read DistrictStats from Databricks and upsert into PostgreSQL.
-
-        Streams rows in batches to stay within the Astro worker memory limit.
-        """
-        params = get_current_context()["params"]
-        skip_watermark = params.get("full_reload") or params.get("full_reload_district_stats")
-        catalog = Variable.get("databricks_catalog")
-        schema = Variable.get("databricks_dbt_schema")
-        batch_size = 5000
-
-        pg_schema = Variable.get("people_api_schema")
-
-        watermark = None
-        if not skip_watermark:
-            with get_postgres_via_ssh() as conn:
-                watermark = get_max_updated_at(conn, pg_schema, "DistrictStats")
-
-        query = (
-            "SELECT "
-            f"{', '.join('to_json(buckets) AS buckets' if c == 'buckets' else c for c in DISTRICT_STATS_COLUMNS)} "
-            f"FROM `{catalog}`.`{schema}`.`m_people_api__districtstats`"
-        )
-        if watermark:
-            query += f" WHERE updated_at >= '{watermark}'"
-            t_log.info("Incremental load — watermark: %s", watermark)
-        elif skip_watermark:
-            t_log.info("Full reload requested — skipping watermark")
-        else:
-            t_log.info("Initial load — table is empty, no watermark")
-        query += " ORDER BY updated_at ASC"
-
-        t_log.info("Reading from Databricks: %s", query)
-        _col_names, batches = read_databricks_table(query, batch_size=batch_size)
-
-        try:
-            with get_postgres_via_ssh() as conn:
-                total = 0
-                for batch in batches:
-                    rows = [
-                        (
-                            row[0],  # district_id
-                            row[1],  # updated_at
-                            row[2],  # total_constituents
-                            row[3],  # total_constituents_with_cell_phone
-                            Json(  # buckets — parse and fix camelCase keys
-                                {_BUCKET_KEY_MAP.get(k, k): v for k, v in json.loads(row[4]).items()}
-                            ),
-                        )
-                        for row in batch
-                    ]
-                    total += upsert_rows(
-                        conn=conn,
-                        schema=pg_schema,
-                        table="DistrictStats",
-                        columns=DISTRICT_STATS_COLUMNS,
-                        conflict_columns=["district_id"],
-                        rows=rows,
-                    )
-        finally:
-            batches.close()
-
-        t_log.info("Upserted %d DistrictStats rows to PostgreSQL", total)
-
-    load_districts() >> load_district_stats()
+    # dbt gate must pass before unload; unload + provision then run in parallel and both feed
+    # create-schema; then the serial load chain.
+    inspect_prod >> dbt_test_voter_gate >> [unload, provision]
+    [unload, provision] >> create_schema
+    create_schema >> copy >> build_indexes >> resize >> validate
 
 
 load_people_api()
